Listener hinzugefügt

+++ MoerderCode/app/src/main/java/com/example/jendrik/moerder/GameObjekts/Clue.java
@@ -6,10 +6,8 @@ import com.google.firebase.database.IgnoreExtraProperties;
 
 import java.io.Serializable;
 
-import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import lombok.RequiredArgsConstructor;
 import lombok.Setter;
 
 /**
@@ -22,9 +20,9 @@ import lombok.Setter;
 public class Clue implements Serializable {
 
     private String name;
-    private double id; //0=player, 1=room, 2=weapon --> für Bildzuweisung
+    private Double id; //0=player, 1=room, 2=weapon --> für Bildzuweisung
 
-    public Clue(String name, int id) {
+    public Clue(String name, Double id) {
         this.name = name;
         this.id = id;
     }
